refactor: move findModeCollisions into the shared DTCG library
sd-native.mjs needs the mode-collision check and must not import the
validator. lib/dtcg.mjs is where flattenDtcg and resolveValue already
live for the same reason. validate-token-output.mjs re-exports it, so
its public surface and test file are unchanged.
