+++ scripts/lib/dtcg.test.mjs
@@ -1,6 +1,6 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import { flattenDtcg, resolveValue } from './dtcg.mjs';
+import { flattenDtcg, resolveValue, findModeCollisions } from './dtcg.mjs';
 
 // text.sm is a DUAL-NODE token: it carries its own $value AND a child.
 const dtcg = {
@@ -55,4 +55,41 @@ test('resolveValue throws on a missing token', () => {
 
 test('resolveValue throws on a circular reference', () => {
   assert.throws(() => resolveValue('a', { a: '{b}', b: '{a}' }), /circular/);
+});
+
+test('findModeCollisions reports a path defined differently across files', () => {
+  const collisions = findModeCollisions([
+    { file: 'light.json', dtcg: { color: { bg: { $value: '#ffffff' } } } },
+    { file: 'dark.json', dtcg: { color: { bg: { $value: '#000000' } } } },
+  ]);
+  assert.equal(collisions.length, 1);
+  assert.equal(collisions[0].path, 'color.bg');
+  assert.deepEqual(
+    collisions[0].defs.map((d) => d.file),
+    ['light.json', 'dark.json'],
+  );
+});
+
+test('findModeCollisions ignores a path repeated with an identical value', () => {
+  const collisions = findModeCollisions([
+    { file: 'a.json', dtcg: { color: { bg: { $value: '#ffffff' } } } },
+    { file: 'b.json', dtcg: { color: { bg: { $value: '#ffffff' } } } },
+  ]);
+  assert.deepEqual(collisions, []);
+});
+
+test('findModeCollisions returns empty for a single source', () => {
+  const collisions = findModeCollisions([
+    { file: 'only.json', dtcg: { color: { bg: { $value: '#ffffff' } } } },
+  ]);
+  assert.deepEqual(collisions, []);
+});
+
+test('findModeCollisions sees dual-node children', () => {
+  const collisions = findModeCollisions([
+    { file: 'a.json', dtcg: { text: { sm: { $value: '14px', lineHeight: { $value: '20px' } } } } },
+    { file: 'b.json', dtcg: { text: { sm: { $value: '14px', lineHeight: { $value: '24px' } } } } },
+  ]);
+  assert.equal(collisions.length, 1);
+  assert.equal(collisions[0].path, 'text.sm.lineHeight');
 });
